+++ pyramids/__init__.py
@@ -1,3 +1,4 @@
+"""pyramids - GIS utility package."""
 from pyramids.config import Config
 
 __all__ = ["dataset", "netcdf", "featurecollection"]