+++ tests/dataset/test_datacube.py
@@ -7,7 +7,8 @@ import geopandas as gpd
 import numpy as np
 from osgeo import gdal
 
-from pyramids.dataset import Datacube, Dataset
+from pyramids.dataset import Dataset
+from pyramids.datacube import Datacube
 
 
 class TestCreateDataCube: