+++ HISTORY.rst
@@ -173,6 +173,44 @@ of the real values in the bands.
 
 0.7.0 (2024-06-01)
 ------------------
+* install viz, dev, all packages through pip.
 * create a separate module for the netcdf files.
 * add configuration file and module for setting gdal configurations.
-* install viz, dev, all packages through pip.
+
+AbstractDataset
+"""""""
+* add `meta_data` property to return the metadata of the dataset.
+* add `access` property to indicate the access mode of the dataset.
+
+
+Dataset
+"""""""
+* add extra parameter `file_i` to the `read_file` method to read a specific file in a compressed file.
+* initialize the `GDAL_TIFF_INTERNAL_MASK` configuration to `No`
+* the add the `access` parameter to the constructor to set the access mode of the dataset.
+* add the `band_units` property to return the units of the bands.
+* the `__str__` and the `__repr__` methods return string numpy like data type (instead of the gdal constant) of the
+dataset.
+* add `meta_data` property setter to set any key:value as a metadata of the dataset.
+* add `scale` and `offset` properties to set the scale and offset of the bands.
+* add `copy` method to copy the dataset to memory.
+* add `get_attribute_table`set_attribute_table` method to get/set the attribute table of a specific band.
+* the `plot` method uses the rgb bands defined in the dataset plotting (if exist).
+* add `create` method to create a new dataset from scratch.
+* add `write_array` method to write an array to an existing dataset.
+* add `get_mask` method to get the mask of a dataset band.
+* add `band_color` method to get the color assigned to a specific band (RGB).
+* add `get_band_by_color` method to get the band index by its color.
+* add `get_histogram` method to get/calculate  the histogram of a specific band.
+* the `read_array` method takes and extra parameter `window` to lazily read a `window` of the raster, the window is
+[xoff, yoff, x-window, y-window], the `window` can also be a geodataframe.
+* add `get_block_arrangement` method divide the raster into tiles based on the block size.
+* add tiff file writing options (compression/tile/tile_length)
+
+Datacube
+""""""""
+* the `Datacube` is moved to a separate module `datacube`.
+
+NetCDF
+"""""""
+* move all the netcdf related functions to a separate module `netcdf`.